Update visualiser config values to doubles
This allow for the use of decimal points for extra precision.

+++ src/main/kotlin/dev/mizarc/bellclaims/infrastructure/persistence/Config.kt
@@ -11,8 +11,8 @@ class Config(val plugin: Plugin) {
     var initialClaimSize = 0
     var minimumPartitionSize = 0
     var distanceBetweenClaims = 0
-    var visualiserHideDelayPeriod = 0
-    var visualiserRefreshPeriod = 0
+    var visualiserHideDelayPeriod = 0.0
+    var visualiserRefreshPeriod = 0.0
     var pluginLanguage = "EN"
 
     init {
@@ -26,8 +26,8 @@ class Config(val plugin: Plugin) {
         initialClaimSize = maxOf(3,configFile.getInt("initial_claim_size"))
         minimumPartitionSize = maxOf(3, configFile.getInt("minimum_partition_size"))
         distanceBetweenClaims = configFile.getInt("distance_between_claims")
-        visualiserHideDelayPeriod = configFile.getInt("visualiser_hide_delay_period")
-        visualiserRefreshPeriod = configFile.getInt("visualiser_refresh_period")
+        visualiserHideDelayPeriod = configFile.getDouble("visualiser_hide_delay_period")
+        visualiserRefreshPeriod = configFile.getDouble("visualiser_refresh_period")
         pluginLanguage = configFile.getString("plugin_language") ?: "EN"
     }
